removed unneeded stuff and renamed appropriate
